Code without contracts
