+++ src/main/kotlin/ru/otus/homework/Address.kt
@@ -1,0 +1,9 @@
+package ru.otus.homework
+
+class Address {
+    lateinit var street1: String
+    lateinit var street2: String
+    lateinit var city: String
+    lateinit var country: String
+    lateinit var index: String
+}
